+++ game-source-code/StopWatch.h
@@ -7,16 +7,16 @@ public:
 	StopWatch() : startTime_(0), elapsedTime_(0), stop_(0) {};
 	void start();
 	void stop();
-	const double getTimeElapsed();
-	const double getStopTime() const;
+	const float getTimeElapsed();
+	const float getStopTime() const;
 
 private:
-	double startTime_;
-	double elapsedTime_;
-	double stop_;
+	float startTime_;
+	float elapsedTime_;
+	float stop_;
 
-	double getProcessTime();
-	double timeElapsed();
+	float getProcessTime();
+	float timeElapsed();
 };
 
 #endif
